feat :: 행동선택 메뉴 캐릭터 정보 뷰 추가

+++ Assets/01_Scripts/Runtime/Battles/CameraControlle/CameraHandler.cs
@@ -56,7 +56,7 @@ public class CameraHandler : Singleton<CameraHandler>
         camera.orthographicSize = targetSize;
     }
     
-    public async Task MoveToLerp(Vector3 targetPosition, int targetSize)
+    public async Task MoveToLerp(Vector3 targetPosition, float targetSize)
     {
         if (isFollowing)
             return;
